pyproject.toml
[project]
name = "mcp-ntopng"
version = "1.0.0"
description = "An MCP server to interact with a the NVD Database API."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.1",
    "clickhouse-driver>=0.2.5",
    "pip-system-certs>=4.0",
    "requests>=2.32.3",
    "fastmcp>=0.4.1",
    "requests>=2.32.3",
]

authors = [{ name = "Graziano Labs Corp." }]
maintainers = [
    { name = "Marco Graziano", email = "marco@graziano.com" },
]
keywords = ["ntop", "mcp", "llm", "cybersecurity"]
license = { text = "MIT" }

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-ntopng = "mcp_ntopng.main:main"
